+++ src/SubscriptionManager.ts
@@ -2,7 +2,7 @@ import type { Observer, Subscription } from "./types.ts";
 
 /**
  * Provides a way to manage subscriptions and notify observers. Used
- * internally by Observable, not meant to be used directly.
+ * internally by Signal, not meant to be used directly.
  */
 export class SubscriptionManager<T> {
   /** Used to call observers in order of subscription. */
@@ -34,6 +34,11 @@ export class SubscriptionManager<T> {
     this.observers.push(observer);
     this.subscriptionsMap.set(observer, subscription);
     return subscription;
+  }
+
+  /** The number of active subscriptions. */
+  get size(): number {
+    return this.subscriptionsMap.size;
   }
 
   /**